Added util_classes unit tests

+++ src/pycubeview/util_classes.py
@@ -5,12 +5,12 @@ from dataclasses import dataclass
 
 @dataclass
 class PixelValue:
-    v: int = 0
-    r: int = 0
-    g: int = 0
-    b: int = 0
+    v: float = 0.0
+    r: float = 0.0
+    g: float = 0.0
+    b: float = 0.0
     pixel_type: Literal["single"] | Literal["rgb"] = "single"
 
     @classmethod
     def null(cls) -> "PixelValue":
-        return cls(-999, -999, -999, -999)
+        return cls(-999.0, -999.0, -999.0, -999.0)
